+++ src/internal/commands/gho.rs
@@ -1,7 +1,7 @@
 use crate::{exec_on, internal::shell::ShellError};
 use anyhow::Context;
 use clap::Args;
-use std::path::Path;
+use std::path::PathBuf;
 use thiserror::Error;
 
 use super::InternalCommandOptions;
@@ -10,19 +10,19 @@ use super::InternalCommandOptions;
 pub struct GitHubOpenOptions {
     /// The path of the project for which to open the Git repository.
     #[arg(default_value = ".")]
-    path: String,
+    path: PathBuf,
 }
 
 #[derive(Error, Debug)]
 pub enum GitHubOpenError {
     #[error("path '{path}' does not exist or is not an accessible directory")]
-    PathDoesNotExist { path: String },
+    PathDoesNotExist { path: PathBuf },
 
     #[error("path '{path}' is not a Git repository")]
-    PathNotAGitRepository { path: String },
+    PathNotAGitRepository { path: PathBuf },
 
     #[error("project at path '{path}' has no remotes configured")]
-    NoRemotesConfigured { path: String },
+    NoRemotesConfigured { path: PathBuf },
 }
 
 pub fn git_hub_open(
@@ -32,14 +32,11 @@ pub fn git_hub_open(
         options,
     }: InternalCommandOptions<GitHubOpenOptions>,
 ) -> anyhow::Result<()> {
-    let path = Path::new(&options.path);
-    let path_str = path.to_string_lossy().to_string();
+    let path = options.path.clone();
+    let path_str = path.to_str().unwrap();
 
     if !path.exists() || !path.is_dir() {
-        return Err(GitHubOpenError::PathDoesNotExist {
-            path: options.path.clone(),
-        }
-        .into());
+        return Err(GitHubOpenError::PathDoesNotExist { path }.into());
     }
 
     let is_git_dir = exec_on!(shell, "git", "-C", path_str, "rev-parse", "--git-dir")
@@ -54,19 +51,19 @@ pub fn git_hub_open(
                 } if status.code() == Some(128) => Ok(false),
                 _ => Err(err),
             },
-            |output| Ok(!output.trim().is_empty()),
+            |output| Ok(!output.stdout.trim().is_empty()),
         )
         .context("checking whether current dir is a git repository")?;
 
     if !is_git_dir {
-        return Err(GitHubOpenError::PathNotAGitRepository { path: path_str }.into());
+        return Err(GitHubOpenError::PathNotAGitRepository { path }.into());
     }
 
     let remote_branches = exec_on!(shell, "git", "-C", path_str, "branch", "-r")
-        .map(|output| output.trim().to_string())?;
+        .map(|output| output.stdout.trim().to_string())?;
 
     if remote_branches.is_empty() {
-        return Err(GitHubOpenError::NoRemotesConfigured { path: path_str }.into());
+        return Err(GitHubOpenError::NoRemotesConfigured { path }.into());
     }
 
     let remote_url = exec_on!(
@@ -78,7 +75,7 @@ pub fn git_hub_open(
         "--get",
         "remote.origin.url"
     )
-    .map(|output| output.trim().to_string())
+    .map(|output| output.stdout.trim().to_string())
     .context("trying to fetch the remote url")?;
 
     exec_on!(shell, "open", remote_url)?;