+++ src/internal/mod.rs
@@ -3,3 +3,4 @@ pub mod commands;
 pub mod constants;
 pub mod progress;
 pub mod shell;
+pub mod utils;