+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## 0.1.2 (2023-03-17)
+
+- feat: add `ford` command
+- refactor: rename `update` command to `upgrade`
+- feat: add `--quiet` flag to all commands (completely disables output)
+- chore: improve verbose logging
+
 ## 0.1.1 (2023-03-16)
 
-- docs: documentation updates and fixes
+- docs: documentation upgrades and fixes
 
 ## 0.1.0 (2023-03-16)